데이터베이스 형식에 맞춰 도메인 멤버 변수 수정
일단 데이터베이스 연결까지 해보고 고민하려고 한 부분이지만
데이터베이스의 컬럼들이 꼭 객체의 멤버 변수로 등록되어 있어야하는지 고민이 필요할 것같다.

+++ src/main/java/kr/codesqaud/cafe/repository/article/ArticleRepository.java
@@ -10,5 +10,5 @@ public interface ArticleRepository {
 
     List<Article> findAllArticles();
 
-    Article findByArticleId(int id);
+    Article findByArticleId(long id);
 }
